Rename package due to PyPI’s naming restrictions.
Publishing resulted in the following error:

```plain
Error: Trusted publishing exchange failure:
Token request failed: the server refused the request for the following reasons:

* `invalid-payload`: The name 'qualify' isn't allowed. See https://pypi.org/help/#project-name for more information.
```

Rename `qualify` to `python-qualify` to accommodate PyPI’s naming
policy.
