+++ debian/changelog
@@ -1,4 +1,4 @@
-qualify (0.1.0) UNRELEASED; urgency=medium
+python3-qualify (0.1.0) UNRELEASED; urgency=medium
 
   * Initial release
