+++ debian/qualify.manpages
@@ -1,1 +1,1 @@
-build/man/qualify.3
+build/man/python_qualify.3
