+++ debian/rules
@@ -3,7 +3,7 @@
 # output every command that modifies files on the build system.
 #export DH_VERBOSE = 1
 
-export PYBUILD_NAME=qualify
+export PYBUILD_NAME=python_qualify
 
 %:
 	dh $@ --with python3 --buildsystem=pybuild
