+++ pyproject.toml
@@ -10,7 +10,7 @@ disallow_subclassing_any = true
 disallow_untyped_calls = true
 disallow_untyped_decorators = true
 disallow_untyped_defs = true
-files = "qualify/**/*.py,tests/**/*.py"
+files = "python_qualify/**/*.py,tests/**/*.py"
 implicit_reexport = false
 no_implicit_optional = true
 strict_equality = true
@@ -19,7 +19,7 @@ warn_return_any = true
 warn_unused_ignores = true
 
 [tool.poetry]
-name = "qualify"
+name = "python-qualify"
 version = "0.1.0"
 description = "Import top-level modules from a directory not on `sys.path`"
 readme = [
@@ -61,11 +61,11 @@ tasks.cmd = "poe -v"
 tasks.help = "List available tasks"
 doc.cmd = "sphinx-build -aqEW -b man doc/sphinx build/man"
 doc.help = "Generate documentation"
-html.cmd = "pdoc qualify !qualify.settings"
+html.cmd = "pdoc python_qualify !python_qualify.settings"
 html.help = "Browse HTML documentation"
-linter.cmd = "pylint --enable-all-extensions qualify"
+linter.cmd = "pylint --enable-all-extensions python_qualify"
 linter.help = "Check for style violations"
-man.cmd = "man build/man/qualify.3"
+man.cmd = "man build/man/python_qualify.3"
 man.help = "Open manual page"
 tests.cmd = "pytest"
 tests.help = "Run test suite"
