+++ python_qualify/__init__.py
@@ -5,10 +5,10 @@
 """
 
 # Re-export these symbols
-# (This promotes them from qualify.qualify to qualify)
-from qualify.api import enable_submodules as enable_submodules
+# (This promotes them from python_qualify.python_qualify to python_qualify)
+from python_qualify.api import enable_submodules as enable_submodules
 
-from qualify.version import version
+from python_qualify.version import version
 
 __all__ = [
     # Tell pdoc to pick up all re-exported symbols
