+++ tests/fixtures/ab/__init__.py
@@ -1,5 +1,5 @@
 """Test package containing modules which refer to each other by
 their unqualified top-level names."""
 
-import qualify
-qualify.enable_submodules(__name__)
+import python_qualify
+python_qualify.enable_submodules(__name__)
